% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition.R
\name{get_data}
\alias{get_data}
\title{Extract the data frame from a cohort object}
\usage{
get_data(cohort)

get_data(cohort)
}
\arguments{
\item{cohort}{A \code{cohort} object.}
}
\value{
The filtered data frame at the current stage.

The filtered data frame at the current stage.
}
\description{
Extract the data frame from a cohort object

Extract the data frame from a cohort object
}
