% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition.R
\name{print.cohort}
\alias{print.cohort}
\title{Print method for cohort objects}
\usage{
\method{print}{cohort}(x, ...)

\method{print}{cohort}(x, ...)
}
\arguments{
\item{x}{A \code{cohort} object.}

\item{...}{Ignored.}
}
\description{
Print method for cohort objects

Print method for cohort objects
}
