% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition.R
\name{get_data}
\alias{get_data}
\title{Extract the data frame from a cohort object}
\usage{
get_data(cohort)
}
\arguments{
\item{cohort}{A \code{cohort} object.}
}
\value{
The filtered data frame at the current stage.
}
\description{
Extract the data frame from a cohort object
}
\examples{
cgd <- prep_cgd_example()
cohort <- new_cohort(cgd, label = "Randomised", id_col = "id") |>
  include_if(age >= 5, "Age >= 5 years")
get_data(cohort)

}
