% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition.R
\name{get_tracker}
\alias{get_tracker}
\title{Extract the attrition tracker from a cohort object}
\usage{
get_tracker(cohort)
}
\arguments{
\item{cohort}{A \code{cohort} object.}
}
\value{
A tibble with columns \code{group}, \code{step}, \code{n_remaining}, \code{n_dropped}.
}
\description{
Extract the attrition tracker from a cohort object
}
\examples{
cgd <- prep_cgd_example()
cohort <- new_cohort(cgd, label = "Randomised", id_col = "id") |>
  include_if(age >= 5, "Age >= 5 years")
get_tracker(cohort)

}
